Revert "API Security using Firebase Authentication."
