+++ TruckBooking/.gitignore
@@ -32,4 +32,3 @@ build/
 ### VS Code ###
 .vscode/
 /bin/
-
