+++ TruckBooking/pom.xml
@@ -80,7 +80,11 @@
 			<groupId>javax.xml.bind</groupId>
 			<artifactId>jaxb-api</artifactId>
 		</dependency>
-
+		<dependency>
+			<groupId>com.google.firebase</groupId>
+			<artifactId>firebase-admin</artifactId>
+			<version>8.0.0</version>
+		</dependency>
 		<!-- https://mvnrepository.com/artifact/io.rest-assured/rest-assured -->
 
 
@@ -157,12 +161,6 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-mail</artifactId>
 		</dependency>
-
-		<dependency>
-			<groupId>com.google.firebase</groupId>
-			<artifactId>firebase-admin</artifactId>
-			<version>6.13.0</version>
-		</dependency>
 	</dependencies>
 
 	<dependencyManagement>
@@ -172,7 +170,6 @@
 				<artifactId>springfox-swagger2</artifactId>
 				<version>3.0.0</version>
 			</dependency>
-
 		</dependencies>
 	</dependencyManagement>
