+++ TruckBooking/src/main/java/com/TruckBooking/Booking/Entities/BookingData.java
@@ -55,7 +55,6 @@ public @Data class BookingData {
 
 	private Long rate;
 	private Unit unitValue;
-
 	private String lr;            // optional
 	private String remarks;       // optional
 	private Integer damage;        // optional
@@ -73,4 +72,7 @@ public @Data class BookingData {
 
 	@CreationTimestamp
 	public Timestamp timestamp;
+
+
+
 }
