+++ TruckBooking/src/main/java/com/TruckBooking/LoadsApi/Config/FirebaseConfig.java
@@ -1,0 +1,28 @@
+package com.TruckBooking.LoadsApi.Config;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+import org.springframework.context.annotation.Configuration;
+
+import com.google.auth.oauth2.GoogleCredentials;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.FirebaseOptions;
+
+@Configuration
+public class FirebaseConfig {
+
+	//	@PostConstruct
+	public void Initialize() throws IOException {
+
+		InputStream serviceAccount =this.getClass().getClassLoader().getResourceAsStream("./firebase.json");
+
+		FirebaseOptions options = FirebaseOptions.builder()
+				.setCredentials(GoogleCredentials.fromStream(serviceAccount))
+				.build();
+
+		FirebaseApp Myapp =FirebaseApp.initializeApp(options);
+
+		//		System.out.println(Myapp.getName());
+	}
+}
