+++ TruckBooking/src/main/java/com/TruckBooking/LoadsApi/Config/SecurityConfig.java
@@ -1,0 +1,24 @@
+package com.TruckBooking.LoadsApi.Config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig extends WebSecurityConfigurerAdapter{
+
+	@Override
+	protected void configure(HttpSecurity http) throws Exception {
+		http.csrf().disable()
+		// dont authenticate this particular request
+		.authorizeRequests().antMatchers("/**").permitAll()
+		.and()
+		// make sure we use stateless session; session won't be used to store user's state.
+		.sessionManagement()
+		.sessionCreationPolicy(SessionCreationPolicy.STATELESS);
+	}
+
+}
