+++ TruckBooking/src/main/resources/application.properties
@@ -1,10 +1,12 @@
 #spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
+
+#spring.datasource.url=jdbc:postgresql://10.2.2.10:5432/postgres?useSSL=false
+#spring.datasource.url=jdbc:postgresql://docker-postgres:5432/postgres?useSSL=false
+
 #spring.datasource.url=jdbc:postgresql://18.189.50.49:5432/postgres?useSSL=false
 #spring.datasource.url=jdbc:postgresql://172.31.0.61:5432/postgres?useSSL=false
 #spring.datasource.url=jdbc:postgresql://172.31.7.41:5432/postgres?useSSL=false
 #spring.datasource.url=jdbc:postgresql://172.31.10.179:5432/postgres?useSSL=false
-
-
 
 server.port = 8080
 spring.jpa.database = POSTGRESQL
@@ -42,3 +44,4 @@ spring.servlet.multipart.max-request-size=3MB
 ##spring.jpa.hibernate.ddl-auto=create-drop
 #spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
 spring.mvc.pathmatch.matching-strategy = ANT_PATH_MATCHER
+
